refactor: extract OwidGdocHeader components to their own files
