+++ site/owid.scss
@@ -80,6 +80,7 @@
 @import "./DataValue.scss";
 @import "./AnnotatingDataValue.scss";
 @import "./gdocs/centered-article.scss";
+@import "./gdocs/topic-page.scss";
 @import "./gdocs/SDGGrid.scss";
 @import "./gdocs/SDGTableOfContents.scss";
 @import "./gdocs/MissingData.scss";
